+++ qft/duostra/include/q_device.hpp
@@ -23,12 +23,12 @@ class Operation {
     friend std::ostream& operator<<(std::ostream&, const Operation&);
     friend void to_json(json& j, const Operation& op);
     Operation(Operator oper,
-              std::tuple<unsigned, unsigned> qs,
-              std::tuple<unsigned, unsigned> du)
+              std::tuple<size_t, size_t> qs,
+              std::tuple<size_t, size_t> du)
         : oper_(oper), qubits_(qs), duration_(du) {
         // sort qs
-        unsigned a = std::get<0>(qs);
-        unsigned b = std::get<1>(qs);
+        size_t a = std::get<0>(qs);
+        size_t b = std::get<1>(qs);
         assert(a != b);
         if (a > b) {
             qubits_ = std::make_tuple(b, a);
@@ -46,17 +46,17 @@ class Operation {
         return *this;
     }
 
-    unsigned get_cost() const { return std::get<1>(duration_); }
-    unsigned get_op_time() const { return std::get<0>(duration_); }
-    std::tuple<unsigned, unsigned> get_duration() const { return duration_; }
+    size_t get_cost() const { return std::get<1>(duration_); }
+    size_t get_op_time() const { return std::get<0>(duration_); }
+    std::tuple<size_t, size_t> get_duration() const { return duration_; }
     Operator get_operator() const { return oper_; }
     std::string get_operator_name() const { return operator_get_name(oper_); }
-    std::tuple<unsigned, unsigned> get_qubits() const { return qubits_; }
+    std::tuple<size_t, size_t> get_qubits() const { return qubits_; }
 
    private:
     Operator oper_;
-    std::tuple<unsigned, unsigned> qubits_;
-    std::tuple<unsigned, unsigned> duration_;  // <from, to>
+    std::tuple<size_t, size_t> qubits_;
+    std::tuple<size_t, size_t> duration_;  // <from, to>
 };
 
 std::ostream& operator<<(std::ostream&, Operation&);
@@ -67,7 +67,7 @@ bool op_order(const Operation& a, const Operation& b);
 class AStarNode {
    public:
     friend class AStarComp;
-    AStarNode(unsigned cost, unsigned id, bool swtch)
+    AStarNode(size_t cost, size_t id, bool swtch)
         : est_cost_(cost), id_(id), swtch_(swtch) {}
     AStarNode(const AStarNode& other)
         : est_cost_(other.est_cost_), id_(other.id_), swtch_(other.swtch_) {}
@@ -80,12 +80,12 @@ class AStarNode {
     }
 
     bool get_swtch() const { return swtch_; }
-    unsigned get_id() const { return id_; }
-    unsigned get_cost() const { return est_cost_; }
+    size_t get_id() const { return id_; }
+    size_t get_cost() const { return est_cost_; }
 
    private:
-    unsigned est_cost_;
-    unsigned id_;
+    size_t est_cost_;
+    size_t id_;
     bool swtch_;  // false q0 propagate, true q1 propagate
 };
 
@@ -99,42 +99,42 @@ class AStarComp {
 class Qubit {
    public:
     friend std::ostream& operator<<(std::ostream& os, const device::Qubit& q);
-    Qubit(const unsigned i);
+    Qubit(const size_t i);
     Qubit(const Qubit& other);
     Qubit(Qubit&& other);
 
-    unsigned get_id() const;
-    unsigned get_avail_time() const;
+    size_t get_id() const;
+    size_t get_avail_time() const;
     bool is_adj(const Qubit& other) const;
-    unsigned get_topo_qubit() const;
+    size_t get_topo_qubit() const;
 
-    void add_adj(unsigned i);
-    void set_topo_qubit(unsigned i);
-    void set_occupied_time(unsigned t);
-    const std::vector<unsigned>& get_adj_list() const;
+    void add_adj(size_t i);
+    void set_topo_qubit(size_t i);
+    void set_occupied_time(size_t t);
+    const std::vector<size_t>& get_adj_list() const;
 
     // A*
-    unsigned get_cost() const;
+    size_t get_cost() const;
     bool is_marked() const;
     bool is_taken() const;
     bool get_swtch() const;
-    unsigned get_pred() const;
-    unsigned get_swap_time() const;
-    void mark(bool swtch, unsigned pred);
+    size_t get_pred() const;
+    size_t get_swap_time() const;
+    void mark(bool swtch, size_t pred);
     void reset();
-    void take_route(unsigned cost, unsigned swap_time);
+    void take_route(size_t cost, size_t swap_time);
 
    private:
-    unsigned id_;
-    std::vector<unsigned> adj_list_;
-    unsigned topo_qubit_;
-    unsigned occupied_until_;
+    size_t id_;
+    std::vector<size_t> adj_list_;
+    size_t topo_qubit_;
+    size_t occupied_until_;
 
     // for A*
     bool marked_;
-    unsigned pred_;
-    unsigned cost_;
-    unsigned swap_time_;
+    size_t pred_;
+    size_t cost_;
+    size_t swap_time_;
     bool swtch_;
     bool taken_;
 };
@@ -143,40 +143,40 @@ std::ostream& operator<<(std::ostream& os, const device::Qubit& q);
 
 class Device {
    public:
-    Device(std::fstream& file, unsigned r, unsigned s, unsigned cx) noexcept;
-    Device(std::vector<std::vector<unsigned>>&,
-           unsigned r,
-           unsigned s,
-           unsigned cx) noexcept;
+    Device(std::fstream& file, size_t r, size_t s, size_t cx) noexcept;
+    Device(std::vector<std::vector<size_t>>&,
+           size_t r,
+           size_t s,
+           size_t cx) noexcept;
     Device(const Device& other) noexcept;
     Device(Device&& other) noexcept;
 
-    const unsigned SINGLE_CYCLE, SWAP_CYCLE, CX_CYCLE;
+    const size_t SINGLE_CYCLE, SWAP_CYCLE, CX_CYCLE;
 
-    unsigned get_num_qubits() const;
-    std::vector<unsigned> mapping() const;
-    unsigned get_shortest_cost(unsigned i, unsigned j) const;
+    size_t get_num_qubits() const;
+    std::vector<size_t> mapping() const;
+    size_t get_shortest_cost(size_t i, size_t j) const;
 
-    Qubit& get_qubit(unsigned i);
-    const Qubit& get_qubit(unsigned i) const;
+    Qubit& get_qubit(size_t i);
+    const Qubit& get_qubit(size_t i) const;
 
-    Operation execute_single(Operator op, unsigned q);
+    Operation execute_single(Operator op, size_t q);
     std::vector<Operation> duostra_routing(Operator op,
-                                           std::tuple<unsigned, unsigned> qs,
+                                           std::tuple<size_t, size_t> qs,
                                            bool orient);
     std::vector<Operation> apsp_routing(Operator op,
-                                        std::tuple<unsigned, unsigned> qs,
+                                        std::tuple<size_t, size_t> qs,
                                         bool orient);
 
     void print_device_state(std::ostream& out);
-    void place(std::vector<unsigned>& assign);  // topo2device
+    void place(std::vector<size_t>& assign);  // topo2device
     std::vector<Operation>& get_operations();
     void init_apsp();
     void reset();
 
    private:
     // A*
-    std::tuple<bool, unsigned> touch_adj(
+    std::tuple<bool, size_t> touch_adj(
         device::Qubit& qubit,
         std::priority_queue<device::AStarNode,
                             std::vector<device::AStarNode>,
@@ -190,14 +190,14 @@ class Device {
                                      device::Qubit& t1);  // standalone
 
     // apsp
-    std::tuple<unsigned, unsigned> next_swap_cost(unsigned source,
-                                                  unsigned target);
+    std::tuple<size_t, size_t> next_swap_cost(size_t source,
+                                                  size_t target);
 
     // general
     void apply_gate(const Operation& op);
 
     // data member
     std::vector<Qubit> qubits_;
-    std::vector<std::vector<unsigned>> shortest_path_, shortest_cost_;
+    std::vector<std::vector<size_t>> shortest_path_, shortest_cost_;
 };
 }  // namespace device
