+++ qft/duostra/src/schedulers/onion.cpp
@@ -19,11 +19,11 @@ void Onion::assign_gates(unique_ptr<QFTRouter> router) {
     auto gen_to_gates = first_mode_ ? topo_->gate_by_dist_to_first()
                                     : topo_->gate_by_dist_to_last();
 
-    unsigned num_gates = topo_->get_num_gates();
+    size_t num_gates = topo_->get_num_gates();
 
     Tqdm bar{num_gates};
 
-    // unsigned total_size = 0;
+    // size_t total_size = 0;
     // for (const auto& gg : gen_to_gates) {
     //     total_size += gg.second.size();
     // }
@@ -33,20 +33,20 @@ void Onion::assign_gates(unique_ptr<QFTRouter> router) {
         auto youngest =
             conf_.cost_typ
                 ? max_element(gen_to_gates.begin(), gen_to_gates.end(),
-                              [](const pair<unsigned, vector<unsigned>>& a,
-                                 const pair<unsigned, vector<unsigned>>& b) {
+                              [](const pair<size_t, vector<size_t>>& a,
+                                 const pair<size_t, vector<size_t>>& b) {
                                   return a.first < b.first;
                               })
                 : min_element(gen_to_gates.begin(), gen_to_gates.end(),
-                              [](const pair<unsigned, vector<unsigned>>& a,
-                                 const pair<unsigned, vector<unsigned>>& b) {
+                              [](const pair<size_t, vector<size_t>>& a,
+                                 const pair<size_t, vector<size_t>>& b) {
                                   return a.first < b.first;
                               });
 
         auto& wait_list = youngest->second;
         const auto wait_list_size = wait_list.size();
         for (size_t jj = 0; jj < wait_list_size; ++jj, bar.add()) {
-            unsigned gate_idx = get_executable(*router, wait_list);
+            size_t gate_idx = get_executable(*router, wait_list);
             gate_idx = greedy_fallback(*router, wait_list, gate_idx);
 
             auto erase_idx =
